make thermal's device impossible to pick up with carry on mod
Item Allocator and Vacuumulator have the feature of filtering, that is accesible with shift-right-clicking them. However, doing that in-game actually picks them up, so accessing this filtering is impossible.

This change adds Thermal Expansion's devices to Carry On blacklist to "unlock" this behavior again. This list is not meta-sensitive, so it blocks all devices and not just allocator and vacuumulator (which is unfortunate).