fix: 📖Prevent Loot Chests to tier up
The combination of loot boxes was originally done in order to offer players an option that they should have rejected as irrational. But experience has shown that in order to collect a mythical loot box, players were willing to save up ~ 6,000 "Regular" boxes, despite the fact that this is simply not possible.

In fact, hoarding loot boxes is the most frustrating decision you could make. Converting 1 Mythic into ~700 "Unusual" or 81 "Rare" ones is much more profitable and actually more useful to speed up the game.

Loot boxes still require better reworking, but for now, it will be impossible to save them for the sake of an upgrade.